+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.6.4 (23 December, 2024)
+
+- Improvements triggered by #50: 'Add favourite files to "Go to File…" palette'
+  - Added `Favorites: Quick pick` command
+  - Added user prompt if detected that VSCode have preview mode enabled which impacts the UX of opening the files from the favorites list. The prompt has option to fix the incompatible setting.
+
 ## 1.6.3 (15 September, 2024)
 
 - Issue #49: Version 1.6.0 adds the [object Object] instead of the filename